refactor: 메모 리스트 클라이언트 구조 정리
