+++ gridam/src/app/(main)/memo/page.tsx
@@ -1,9 +1,65 @@
+import Link from 'next/link'
 import { getMemoListAction } from '@/features/memo/api/memo.action'
-import MemoListPage from '@/features/memo/components/memo-list-page'
+import { Card } from '@/shared/ui/card'
+import TagBadge from '@/shared/ui/tagbadge'
+import MemoCreateButton from '@/features/memo/components/memo-create-button'
 
 export default async function Page() {
   const { ok, data } = await getMemoListAction()
   const memos = ok ? data : []
+  const hasMemos = memos.length > 0
 
-  return <MemoListPage memos={memos} />
+  return (
+    <div className="mx-auto flex min-h-[calc(100vh-120px)] max-w-2xl flex-col gap-6 px-4 py-10">
+      <header className="flex items-center justify-between">
+        <div className="flex flex-col gap-1">
+          <h1 className="text-2xl font-bold text-foreground">메모장</h1>
+          <p className="text-sm text-muted-foreground">자유롭게 메모를 작성해보세요.</p>
+        </div>
+
+        <MemoCreateButton />
+      </header>
+
+      {hasMemos ? (
+        <section className="flex flex-col gap-3">
+          {memos.map((memo) => (
+            <Link key={memo.id} href={`/memo/${memo.id}`}>
+              <Card
+                className="
+                  cursor-pointer rounded-3xl bg-card px-6 py-4
+                  shadow-sm transition hover:shadow-md
+                "
+              >
+                <div className="flex items-center justify-between gap-4">
+                  <span className="truncate text-base font-medium text-foreground">
+                    {memo.title || '(제목 없음)'}
+                  </span>
+                  <span className="shrink-0 text-xs text-muted-foreground">
+                    {new Date(memo.created_at).toLocaleDateString().replace(/\./g, '.').trim()}
+                  </span>
+                </div>
+
+                {memo.tags && memo.tags.length > 0 && (
+                  <div className="mt-2 flex flex-wrap gap-2">
+                    {memo.tags.map((tag) => (
+                      <TagBadge key={tag}>{tag}</TagBadge>
+                    ))}
+                  </div>
+                )}
+              </Card>
+            </Link>
+          ))}
+        </section>
+      ) : (
+        <section className="flex flex-1 flex-col items-center justify-center gap-4 text-center">
+          <p className="text-sm text-muted-foreground">
+            아직 메모가 없어요.
+            <br />
+            오른쪽 상단의 <span className="font-semibold">+ 새 메모</span> 버튼을 눌러 첫 메모를
+            작성해보세요.
+          </p>
+        </section>
+      )}
+    </div>
+  )
 }
